added jsonwebtoken packages

+++ Server/Controllers/movie.js
@@ -3,81 +3,78 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.DeleteMovie = exports.UpdateMovie = exports.AddMovie = exports.DisplayMovieByID = exports.DisplayMovieList = void 0;
+exports.DeleteMovie = exports.UpdateMovie = exports.AddMovie = exports.DisplayMovieById = exports.DisplayMovieList = void 0;
 const movie_1 = __importDefault(require("../Models/movie"));
-function SanitizeArray(unsanitizedArray) {
-    let sanitizedArray = Array();
-    for (const unsanitizedString of unsanitizedArray) {
-        sanitizedArray.push(unsanitizedString.trim());
-    }
-    return sanitizedArray;
-}
+const Util_1 = require("../Util");
 function DisplayMovieList(req, res, next) {
     movie_1.default.find({})
-        .then(function (data) {
-        res.status(200).json({ success: true, msg: "Movie List Retrieved and Displayed", data: data });
+        .then((data) => {
+        res.status(200).json({ success: true, msg: "Movie List Retrieved and Displayed", data: data, token: null });
     })
-        .catch(function (err) {
+        .catch((err) => {
         console.error(err);
     });
 }
 exports.DisplayMovieList = DisplayMovieList;
-function DisplayMovieByID(req, res, next) {
-    try {
-        let id = req.params.id;
+function DisplayMovieById(req, res, next) {
+    let id = req.params.id;
+    if (id.length != 24) {
+        res.status(400).json({ success: false, msg: "A valid ID is required to retrieve a movie", data: null, token: null });
+    }
+    else {
         movie_1.default.findById({ _id: id })
-            .then(function (data) {
-            res.status(200).json({ success: true, msg: "Move Retrieved by ID", data: data });
+            .then((data) => {
+            if (data) {
+                res.status(200).json({ success: true, msg: "One Movie Retrieved and Displayed", data: data, token: null });
+            }
+            else {
+                res.status(404).json({ success: false, msg: "Movie not found", data: null, token: null });
+            }
         })
-            .catch(function (err) {
+            .catch((err) => {
             console.error(err);
         });
-    }
-    catch {
-        res.json({ success: false, msg: "No Data to Display" });
     }
 }
-exports.DisplayMovieByID = DisplayMovieByID;
+exports.DisplayMovieById = DisplayMovieById;
 function AddMovie(req, res, next) {
-    try {
-        let genres = SanitizeArray(req.body.genres.split(","));
-        let directors = SanitizeArray(req.body.directors.split(","));
-        let writers = SanitizeArray(req.body.writers.split(","));
-        let actors = SanitizeArray(req.body.actors.split(","));
-        let movie = new movie_1.default({
-            movieID: req.body.movieID,
-            title: req.body.title,
-            studio: req.body.studio,
-            genres: genres,
-            directors: directors,
-            writers: writers,
-            actors: actors,
-            length: req.body.length,
-            year: req.body.year,
-            shortDescription: req.body.shortDescription,
-            mpaRating: req.body.mpaRating,
-            criticsRating: req.body.criticsRating
-        });
-        movie_1.default.create(movie)
-            .then(function () {
-            res.status(200).json({ success: true, msg: "Movie Added Successfully", data: movie });
-        })
-            .catch(function (err) {
-            console.error(err);
-        });
-    }
-    catch {
-        res.json({ success: false, msg: "No Data to Add" });
-    }
+    let genres = (req.body.genres) ? (0, Util_1.SanitizeArray)(req.body.genres) : (0, Util_1.SanitizeArray)("");
+    let directors = (req.body.directors) ? (0, Util_1.SanitizeArray)(req.body.directors) : (0, Util_1.SanitizeArray)("");
+    let actors = (req.body.actors) ? (0, Util_1.SanitizeArray)(req.body.actors) : (0, Util_1.SanitizeArray)("");
+    let writers = (req.body.writers) ? (0, Util_1.SanitizeArray)(req.body.writers) : (0, Util_1.SanitizeArray)("");
+    let movie = new movie_1.default({
+        movieID: req.body.movieID,
+        title: req.body.title,
+        studio: req.body.studio,
+        genres: genres,
+        directors: directors,
+        writers: writers,
+        actors: actors,
+        length: req.body.length,
+        year: req.body.year,
+        shortDescription: req.body.shortDescription,
+        mpaRating: req.body.mpaRating,
+        criticsRating: req.body.criticsRating
+    });
+    movie_1.default.create(movie)
+        .then(() => {
+        res.status(200).json({ success: true, msg: "Movie added", data: movie, token: null });
+    })
+        .catch((err) => {
+        console.error(err);
+    });
 }
 exports.AddMovie = AddMovie;
 function UpdateMovie(req, res, next) {
-    try {
-        let id = req.params.id;
-        let genres = SanitizeArray(req.body.genres.split(","));
-        let directors = SanitizeArray(req.body.directors.split(","));
-        let writers = SanitizeArray(req.body.writers.split(","));
-        let actors = SanitizeArray(req.body.actors.split(","));
+    let id = req.params.id;
+    if (id.length != 24) {
+        res.status(400).json({ success: false, msg: "A valid ID is required to update a movie", data: null, token: null });
+    }
+    else {
+        let genres = (req.body.genres) ? (0, Util_1.SanitizeArray)(req.body.genres) : (0, Util_1.SanitizeArray)("");
+        let directors = (req.body.directors) ? (0, Util_1.SanitizeArray)(req.body.directors) : (0, Util_1.SanitizeArray)("");
+        let actors = (req.body.actors) ? (0, Util_1.SanitizeArray)(req.body.actors) : (0, Util_1.SanitizeArray)("");
+        let writers = (req.body.writers) ? (0, Util_1.SanitizeArray)(req.body.writers) : (0, Util_1.SanitizeArray)("");
         let movieToUpdate = new movie_1.default({
             _id: id,
             movieID: req.body.movieID,
@@ -93,32 +90,30 @@ function UpdateMovie(req, res, next) {
             mpaRating: req.body.mpaRating,
             criticsRating: req.body.criticsRating
         });
+        console.log(movieToUpdate);
         movie_1.default.updateOne({ _id: id }, movieToUpdate)
-            .then(function () {
-            res.status(200).json({ success: true, msg: "Movie Updated Successfully", data: movieToUpdate });
+            .then(() => {
+            res.status(200).json({ success: true, msg: "Movie updated", data: movieToUpdate, token: null });
         })
-            .catch(function (err) {
+            .catch((err) => {
             console.error(err);
         });
-    }
-    catch {
-        res.json({ success: false, msg: "No Data to Update" });
     }
 }
 exports.UpdateMovie = UpdateMovie;
 function DeleteMovie(req, res, next) {
-    try {
-        let id = req.params.id;
+    let id = req.params.id;
+    if (id.length != 24) {
+        res.status(400).json({ success: false, msg: "A valid ID is required to delete a movie", data: null, token: null });
+    }
+    else {
         movie_1.default.deleteOne({ _id: id })
-            .then(function () {
-            res.json(id);
+            .then(() => {
+            res.status(200).json({ success: true, msg: "Movie deleted", data: id, token: null });
         })
-            .catch(function (err) {
+            .catch((err) => {
             console.error(err);
         });
-    }
-    catch {
-        res.json({ success: false, msg: "No Data to Delete" });
     }
 }
 exports.DeleteMovie = DeleteMovie;
